All weapons but axe actually working and a little shop stuff
Couldnt figure out what cooper's goal with the axe was so I just left it as is.
oh also half the shop is implemented.

+++ RoyalFlush/Assets/Scripts/Util/PlayerData.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class PlayerData
 {
     public int level;
-    public int health;
-    public int maxHealth;
+    public float health;
+    public float maxHealth;
     public int gold;
     public float[] position;
     public List<string> dropsLeft;
